+++ php/tests/Application/Command/GlobalFunctionsTest.php
@@ -18,9 +18,9 @@ class GlobalFunctionsTest extends IndexedTest
         $output = $command->getGlobalFunctions();
 
         $this->assertThat($output, $this->arrayHasKey('firstFunction'));
-        $this->assertEquals($output['firstFunction']['fqsen'], 'A\firstFunction');
+        $this->assertEquals($output['firstFunction']['fqcn'], 'A\firstFunction');
         $this->assertThat($output, $this->arrayHasKey('secondFunction'));
-        $this->assertEquals($output['secondFunction']['fqsen'], 'A\secondFunction');
+        $this->assertEquals($output['secondFunction']['fqcn'], 'A\secondFunction');
         $this->assertThat($output, $this->logicalNot($this->arrayHasKey('shouldNotShowUp')));
     }
 }